Show confirm modal with merchant info on configuration before saving new credentials
